+++ internal/store/store.go
@@ -29,14 +29,14 @@ import (
 type StoreLayer interface {
 	CreateRedirect(createdBy, fromPath, toUrl string, startsOn, stopsOn *time.Time) (*Redirect, error)
 	CreateUser(createdBy, email, name, password, role, username string) (*User, error)
-	DeleteRedirect(id uint) error
+	DeleteRedirect(id uint64) error
 	FindActiveRedirectByPath(path string) (*Redirect, error)
 	GetAllRedirects() (*[]Redirect, error)
-	GetRedirect(id uint) (*Redirect, error)
+	GetRedirect(id uint64) (*Redirect, error)
 	GetUser(id uint) (*User, error)
 	GetUserWithUsername(username string) (*User, error)
 	UpdatePassword(id uint, password, updatedBy string) error
-	UpdateRedirect(id uint, updatedBy, fromPath, toUrl string, startsOn, stopsOn *time.Time) (*Redirect, error)
+	UpdateRedirect(id uint64, updatedBy, fromPath, toUrl string, startsOn, stopsOn *time.Time) (*Redirect, error)
 }
 
 type storeLayer struct {
